+++ packages/core/theme/src/index.ts
@@ -1,5 +1,6 @@
 export * from './components/button'
 export * from './components/button-group'
+export * from './components/input'
 export * from './components/spinner'
 
 export * from './layouts/box'