+++ packages/core/theme/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @srcube-taro/theme
 
+## 0.0.3
+
+### Patch Changes
+
+- Components:
+  input
+
 ## 0.0.2
 
 ### Patch Changes