+++ packages/core/ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @srcube-taro/ui
 
+## 0.0.3
+
+### Patch Changes
+
+- Components:
+  input
+- Updated dependencies
+  - @srcube-taro/theme@0.0.3
+  - @srcube-taro/input@0.0.1
+  - @srcube-taro/button@0.0.3
+  - @srcube-taro/layouts@0.0.3
+  - @srcube-taro/spinner@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes